inputSerialCode/pyproject.toml
[tool.poetry]
name = "inputserialcode"
version = "0.1.0"
description = "Input SerialCodes for chara"
authors = ["KAWAI Shun <shun@osstech.co.jp>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
python-dotenv = "^0.14.0"

[tool.poetry.scripts]
main = 'inputserialcode.main:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
